% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_bcs.R
\name{bcs}
\alias{bcs}
\alias{bcs.default}
\alias{as.bcs}
\alias{print.bcs}
\title{Methods for "\code{bcs}" Objects}
\usage{
bcs(object, ...)

\method{bcs}{default}(object, ...)

as.bcs(object)

\method{print}{bcs}(x, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}

\item{x, object}{a character with the abbreviation for a Box-Cox symmetric or log-symmetric 
distribution. See details.}
}
\value{
The \code{bcs} function returns an object of class "\code{bcs}", which consists of a list
    with the following components:
\describe{
  \item{abb}{the abbreviation of the distribution considered in the \code{bcsnsm} package.}
  \item{name}{the corresponding name of the distribution.}
  \item{npar}{number of model parameters.}
  \item{extrap}{logical; if \code{TRUE}, it indicates that the distribution has an extra parameter.}
  \item{start}{a function \code{function(x)} which receives a vector of positive data and returns
       initial values for the parameters of the distribution.}
 }
 
 The "\code{bcs}" objects are used to define the marginal distributions and to estimate the
     parameters of the multivariate BCS-NSM distributions in the \code{bcsnsm} package.
     
 The \code{print()} function returns a basic summary of the BCS distribution in the package,
 with its name, abbreviation, and number of parameters. The \code{as.bcs()} function
 transforms a character with one of the abbreviations shown in the tables above into a
 "\code{bcs}" object.
}
\description{
A "\code{bcs}" object consists of a list with basic information about a 
    Box-Cox symmetric distribution, which is used by other functions within the
    \code{bcsnsm} package.
}
\details{
The class of the Box-Cox symmetric (BCS) distributions was introduced by
    Ferrari and Fumes (2017). It consists of a broad class of probability models for positive
    continuous data, which includes flexible distributions with different levels of skewness
    and tail-heaviness.

    The class includes the Box-Cox \emph{t} (Rigby and Stasinopoulos, 2006),
    Box-Cox normal (or Box-Cox Cole-Green; Cole and Green, 1992), Box-Cox power exponential
    (Rigby and Stasinopoulos, 2004) distributions, and all the class of the
    log-symmetric distributions (Vanegas and Paula, 2016) as special cases.
    
    Let \eqn{Y} be a positive continuous random variable with a BCS distribution
    with parameters \eqn{\mu > 0}, \eqn{\sigma > 0}, and \eqn{\lambda \in \mathbb{R}}
    and density generating function \eqn{r}. The probability density function of
    \eqn{Y} is given by

    \eqn{
    f(y; \mu, \sigma, \lambda) = \left\{\begin{array}{ll}
      \dfrac{y^{\lambda-1}}{\mu^\lambda \sigma} \dfrac{r(z^2)}{R\left(\frac{1}{\sigma |\lambda|}\right)}, & \mbox{ if } \lambda \neq 0,\\
      \dfrac{1}{y\sigma} r(z^2), & \mbox{ if } \lambda = 0,
      \end{array}\right., \quad y > 0,
    }

    where \eqn{r:[0,\infty) \longrightarrow [0, \infty)}
    satisfies \eqn{\int_0^\infty u^{-1/2}r(u)\textrm{d} u = 1},
    \eqn{R(x) = \int_{-\infty}^x r(u^2)\textrm{d} u, x \in \mathbb{R}}, and
    
    \eqn{
    z = \left\{
    \begin{array}{ll}
    \dfrac{1}{\sigma \lambda} \left\{\left(\frac{y}{\mu}\right)^\lambda - 1 \right\}, & \mbox{ if } \lambda \neq 0, \\
    \dfrac{1}{\sigma} \log\left(\frac{y}{\mu}\right), & \mbox{ if } \lambda = 0.
    \end{array}
    \right.
    }

    The function \eqn{r} is called density generating function, and it specifies the
    generating symmetric family of \eqn{Y} within the class of the BCS probability
    models. This function can also depend on extra parameters, such as the Box-Cox \emph{t} and
    Box-Cox power exponential distributions. We call these extra parameters
    \code{zeta}.

    The current available BCS distributions in the \code{bcsnsm} package can be seen below.
  \tabular{llc}{
 \bold{Distribution}  \tab \bold{Abbreviation} \tab \bold{N. of parameters}\cr
 Box-Cox Cauchy  \tab \code{"\link{bcca}"}      \tab  3  \cr
 Box-Cox Hyperbolic  \tab \code{"\link{bchp}"}      \tab  4  \cr
 Box-Cox Laplace  \tab \code{"\link{bcla}"}      \tab  3  \cr
 Box-Cox Type I Logistic  \tab \code{"\link{bcloi}"}      \tab  3  \cr
 Box-Cox Type II Logistic  \tab \code{"\link{bcloii}"}      \tab  3  \cr
 Box-Cox Normal  \tab \code{"\link{bcno}"}      \tab  3  \cr
 Box-Cox Power exponential  \tab \code{"\link{bcpe}"}      \tab  4  \cr
 Box-Cox Slash  \tab \code{"\link{bcsl}"}      \tab  4  \cr
 Box-Cox \emph{t}  \tab \code{"\link{bct}"}      \tab  4  \cr
 }

 Log-symmetric special cases are also available:
  \tabular{llc}{
 \bold{Distribution}  \tab \bold{Abbreviation} \tab \bold{N. of parameters}\cr
 Log-Cauchy  \tab \code{"\link{lca}"}      \tab  2  \cr
Log-Hyperbolic  \tab \code{"\link{lhp}"}      \tab  3  \cr
 Log-Laplace  \tab \code{"\link{lla}"}      \tab  2  \cr
 Log-Type I Logistic  \tab \code{"\link{lloi}"}      \tab  2  \cr
 Log-Type II Logistic  \tab \code{"\link{lloii}"}      \tab  2  \cr
 Log-Normal  \tab \code{"\link{lno}"}      \tab  2  \cr
 Log-Power exponential  \tab \code{"\link{lpe}"}      \tab  3  \cr
Log-Slash  \tab \code{"\link{lsl}"}      \tab  3  \cr
 Log-\emph{t}  \tab \code{"\link{lt}"}      \tab  3  \cr
 }
}
\examples{
bcs("bcno")
bcs("lno")

bcs("bct")
bcs("lt")
}
\references{
Cole, T., and Green, P.J. (1992). Smoothing reference centile curves: the LMS
     method and penalized likelihood. \emph{Statistics in medicine}, 11, 1305-1319.

 Ferrari, S. L., and Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. \emph{AStA Advances in Statistical Analysis}, 101, 321-344.

 Rigby, R. A., and Stasinopoulos, D. M. (2004). Smooth centile curves for skew
     and kurtotic data modelled using the Box-Cox power exponential
     distribution. \emph{Statistics in medicine}, 23, 3053-3076.

 Rigby, R. A., and Stasinopoulos, D. M. (2006). Using the Box-Cox t
     distribution in GAMLSS to model skewness and kurtosis. \emph{Statistical Modelling}, 6, 209-229.

 Vanegas, L. H., and Paula, G. A. (2016). Log-symmetric distributions:
     statistical properties and parameter estimation. \emph{Brazilian Journal of Probability and Statistics}, 30, 196-220.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
