% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_bcsnsm_methods.R
\name{bcsnsm-methods}
\alias{bcsnsm-methods}
\alias{vcov.bcsnsm}
\alias{confint.bcsnsm}
\alias{logLik.bcsnsm}
\alias{AIC.bcsnsm}
\alias{print.bcsnsm}
\alias{summary.bcsnsm}
\alias{print.summary.bcsnsm}
\title{Methods for \code{"bcsnsm"} objects}
\usage{
\method{vcov}{bcsnsm}(object, parm = c("all", "mu", "sigma", "lambda", "nu", "gamma"), ...)

\method{confint}{bcsnsm}(
  object,
  parm = c("all", "mu", "sigma", "lambda", "nu", "gamma"),
  level = 0.95,
  digits = max(3, getOption("digits") - 3),
  ...
)

\method{logLik}{bcsnsm}(object, ...)

\method{AIC}{bcsnsm}(object, ..., k = 2)

\method{print}{bcsnsm}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{bcsnsm}(object, ...)

\method{print}{summary.bcsnsm}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{parm}{character; specifies which submatrix of the asymptotic covariance matrix of the
maximum likelihood estimators should be returned for the \code{vcov} function or
which parameters are to be given confidence intervals for the \code{confint} function. The
options are \code{"all"} (default), \code{"mu"}, \code{"sigma"}, \code{"lambda"}, \code{"nu"}, and \code{"gamma"}.}

\item{...}{additional argument(s) for methods.}

\item{level}{the confidence level required.}

\item{digits}{number of digits in print methods.}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical AIC.}

\item{x, object}{an object of class \code{bcsnsm}.}
}
\description{
Methods for \code{"bcsnsm"} objects
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
