% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_bcs-nsm_distributions.R
\name{bcsnsm.dist}
\alias{bcsnsm.dist}
\alias{dbcsnsm}
\alias{rbcsnsm}
\title{The Class of the Multivariate BCS-NSM Distributions}
\usage{
dbcsnsm(
  x,
  mu,
  sigma,
  lambda,
  nu = NULL,
  Gamma = diag(ncol(x)),
  copula = c("gaussian", "t", "slash", "hyp"),
  eta = NULL,
  margins = "bcno",
  log = FALSE
)

rbcsnsm(
  n,
  mu,
  sigma,
  lambda,
  nu,
  Gamma = diag(ncol(x)),
  copula = c("gaussian", "t", "slash", "hyp"),
  eta = NULL,
  margins = "bcno"
)
}
\arguments{
\item{x}{vector or matrix of non-negative quantiles. If \code{x} is a matrix,
each row is taken to be a quantile.}

\item{mu}{vector of the marginal scale parameters in \code{(0, infty)^d}, where \code{d} is the
dimension.}

\item{sigma}{vector of the marginal relative dispersion parameters in \code{(0, infty)^d}, where \code{d} is the
dimension.}

\item{lambda}{vector of the marginal skewness parameters in \code{R^d}, where \code{d} is the
dimension.}

\item{nu}{vector of possible extra parameters of the marginal distributions. If a particular
marginal distribution has no additional parameter, it must be necessarily specified with
\code{NA} in that position.}

\item{Gamma}{the association matrix. It must be a positive-definite correlation matrix, default is
\code{diag(ncol(x))} (the so-called non-associative association structure).}

\item{copula}{character; informs which normal scale mixture distribution
should be used to generate the NSM copula. Currently,
the copulas available are: Gaussian (\code{"gaussian"}),
Student's t (\code{"t"}), slash (\code{"slash"}), and hyperbolic (\code{"hyp"}).}

\item{eta}{possible extra parameter induced by the copula, parameterized to take values at
\code{(0, 1)}. For instance, under the \emph{t} copula, \code{eta} specifies
\code{eta} / (1 - \code{eta}) degrees of freedom.}

\item{margins}{a character or a character vector; specifies the marginal BCS distributions. If all
BCS margins are the same, it is sufficient to enter only one character. A table with the
current available BCS distributions can be seen in \code{\link[bcnsm]{bcs}}.}

\item{log}{logical; if \code{TRUE}, probabilities \code{p} are given
as \code{log(p)}.}

\item{n}{number of replicates to return.}
}
\value{
\code{dbcsnsm} returns the evaluated joint density function and
     \code{rbcsnsm} generates random values.
}
\description{
Joint probability density function
    and a random generator for the multivariate Box-Cox symmetric
    distributions generated by a normal scale mixture copula (BCS-NSM multivariate distributions).
}
\details{
The BCS-NSM multivariate distributions consists of a broad class of models for positive
    continuous data with Box-Cox symmetric (BCS) marginal distributions, where a normal scale
    mixture (NSM) copula describes the dependence structure.

    Let \eqn{\bold{Y} = (Y_1, \ldots, Y_n)'} be a vector of positive continuous random variables, where
    each \eqn{Y_j} follows a BCS distribution (not necessarily the same) for all
    \eqn{j = 1, \ldots , d}. We say that \eqn{\bold{Y}} follows a \emph{d}-variate
    BCS-NSM distribution if it has an NSM copula with association matrix \eqn{\bold{\Gamma}}
    (\code{Gamma}) and mixing cumulative distribution function (CDF) \eqn{H}.

    The class of the BCS-NSM distributions includes models with various possibilities of marginal
    and dependence settings. For instance, each margin of \eqn{\bold{Y}} can follow a different
    BCS distribution. This property provides the BCS-NSM distributions the flexibility to 
    accommodate random vectors whose marginal distributions have different behaviors, such as 
    different types of skewness and tail-heaviness.

    The NSM copula is characterized by the mixing CDF \eqn{H}. For instance, if \eqn{H} is the
    CDF of a random variable \eqn{W} degenerate at \eqn{w = 1}, then \eqn{\bold{Y}} has a Gaussian
    copula. If \eqn{H} is the CDF of a random variable \eqn{W} with an inverse gamma distribution
    with both parameters equal to \eqn{\delta/2}, \eqn{\delta > 0}, then \eqn{\bold{Y}} has a
    \emph{t} copula with \eqn{\delta} degrees of freedom. In the \code{bcsnsm} package, we consider
    the parameterization \code{eta = delta / (1 + delta)}.
    
    
    Because copulas are invariant under strictly increasing transformations,
    copula-based measures of association of the scale mixtures of normal distributions are the
    same for the BCS-NSM distributions. This statement is valid for measures of concordance, such
    as Kendall's tau or the tail dependence coefficients. Let \eqn{\gamma_{ij}} be the element of
    the \emph{i}th row and \emph{j}th column of the association matrix \eqn{\bold{\Gamma}}, where
    \eqn{\gamma_{ii} = 1}, for \emph{i, j = 1, ..., d}. Thus, Kendall's tau of \eqn{Y_i} and 
    \eqn{Y_j} is given by

    \deqn{\tau(Y_i, Y_j) = \dfrac{2 \; \textrm{asin}(\gamma_{ij})}{\pi},}

    and it is invariant under the copula. Since Kendall’s tau is a
    rank correlation coefficient, it measures any monotonous association between two random variables.
    Thus, we have an increasing monotonous association between \eqn{Y_i} and \eqn{Y_j} when \eqn{\gamma_{ij}}
    is close to \eqn{1}, and a decreasing monotonous association when \eqn{\gamma_{ij}} is close to \eqn{-1}.
    Furthermore, \eqn{\gamma_{ij} = 0} indicates no monotonous association. These facts show that
    the association matrix \eqn{\bold{\Gamma}} plays an essential role in specifying the dependence
    structure of \eqn{\bold{Y}}. It can be parameterized in terms of a real-valued vector \eqn{\gamma}
    in an interpretative way depending on the type of association that is of interest to describe.
}
\examples{
# Sample size and dimension
n <- 500
d <- 4

### Marginal specifications

# Marginal distributions
margins <- c("bct", "bcloi", "bcpe", "bcno")

# Marginal parameters
mu <- c(13, 20, 15, 10)
sigma <- c(0.1, 0.3, 0.4, 0.2)
lambda <- c(-2, 3, 1, -0.5)
nu <- c(2, NA, 8, NA)

### Dependence specifications

# Copula
copula <- "slash"

# Extra parameter
eta <- 0.75

# Association matrix
tau <- 0.6
gamma <- sin(0.5 * pi * tau)
Gamma <- matrix(gamma, d, d)
diag(Gamma) <- 1

### Generating observations
set.seed(1)
x <- rbcsnsm(n, mu, sigma, lambda, nu, Gamma, copula, eta, margins)

### Visualization with mvplot() (see ?mvplot for documentation)
mvplot(x)

}
\references{
Vanegas, L. H., and Paula, G. A. (2016). Log-symmetric distributions:
     statistical properties and parameter estimation. \emph{Brazilian Journal of
     Probability and Statistics}, 30, 196--220.

 Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. \emph{AStA Advances in Statistical Analysis},
     101, 321--344.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
