% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_bcsnsm_classification.R
\name{bcsnsmclass}
\alias{bcsnsmclass}
\title{BCS-NSM Classification}
\usage{
bcsnsmclass(object, test, cl)
}
\arguments{
\item{object}{an object of class \code{"bcsnsmreg"}.}

\item{test}{matrix or data frame of test set cases.}

\item{cl}{factor of true classifications of training set.}
}
\value{
Factor of classifications of test set.
}
\description{
BCS-NSM Classification
}
\examples{
\dontrun{
# Winscosin Breast Cancer Dataset
?wdbc

# Training set index
set.seed(123)
id <- sample(1:nrow(wdbc), 0.7 * nrow(wdbc))

# BCS-NSM regression
fit <- bcsnsmreg(texture + area + smoothness + compactness + concavity ~ diagnosis,
                 data = wdbc[id, ], margins = c("lt", "lt", "lno", "lpe", "bct"))

## Marginal quantile residuals
plot(fit_gaussian, "marginal", panel = c(2, 3))

## Summary
summary(fit_gaussian)

## Overall quantile residuals of the final model
plot(fit_gaussian)

## The epsilon's transformations
plot(fit_gaussian, "epsilon")

## Classification
pred <- bcsnsmclass(fit_gaussian, wdbc[-id, 1:5], wdbc$diagnosis[id])
table(pred = pred, true = wdbc$diagnosis[-id])
}

}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
