% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_regression_methods.R
\name{bcsnsmreg-methods}
\alias{bcsnsmreg-methods}
\alias{model.frame.bcsnsmreg}
\alias{model.matrix.bcsnsmreg}
\alias{vcov.bcsnsmreg}
\alias{confint.bcsnsmreg}
\alias{logLik.bcsnsmreg}
\alias{AIC.bcsnsmreg}
\alias{print.bcsnsmreg}
\alias{summary.bcsnsmreg}
\alias{print.summary.bcsnsmreg}
\title{Methods for "bcsnsmreg" objects}
\usage{
\method{model.frame}{bcsnsmreg}(formula, ...)

\method{model.matrix}{bcsnsmreg}(object, margin, ...)

\method{vcov}{bcsnsmreg}(object, parm = c("all", "beta", "sigma", "lambda", "nu", "gamma"), ...)

\method{confint}{bcsnsmreg}(
  object,
  parm = c("full", "beta", "sigma", "lambda", "nu", "gamma"),
  level = 0.95,
  ...
)

\method{logLik}{bcsnsmreg}(object, ...)

\method{AIC}{bcsnsmreg}(object, ..., k = 2)

\method{print}{bcsnsmreg}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{bcsnsmreg}(object, ...)

\method{print}{summary.bcsnsmreg}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted to each marginal distribution.
For instance, \code{formula = c(y1 + y2 + y3 ~ x1 + x2 | x1 + x3 | x2 + x3)} fits a
3-variate BCS-NSM regression that models \code{y1} as a function of \code{x1} and \code{x2};
\code{y2} as a function of \code{x1} and \code{x3}; and \code{y3} as a function of \code{x2}
and \code{x3}.}

\item{...}{additional argument(s) for methods.}

\item{margin}{an integer which specifies the marginal distribution.}

\item{parm}{a character; specifies which parameters are to be given the asymptotic covariance
matrix of the maximum likelihood estimators (for \code{vcov} function) or the confidence
intervals (for \code{confint} function). The options are \code{"all"} (default), where
all parameters are considered, \code{"beta"}, \code{"sigma"}, \code{"lambda"}, \code{"nu"},
and \code{"gamma"}.}

\item{level}{the confidence level required.}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical Akaike information criteria (AIC).}

\item{x, object}{an object of class \code{"bcsnsmreg"}.}

\item{digits}{a non-null value for \code{digits} specifies the minimum number of
digits to be printed in values.}
}
\description{
Methods for "bcsnsmreg" objects
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
