% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_fit_functions.R
\name{bcsnsmreg}
\alias{bcsnsmreg}
\title{BCS-NSM Regression for Multivariate Positive Data}
\usage{
bcsnsmreg(
  formula,
  data,
  subset,
  na.action,
  margins = "bcno",
  links = "log",
  association = c("unstructured", "uniform", "nonassociative"),
  copula = c("gaussian", "t", "slash", "hyp"),
  eta = NULL,
  control = control_fit(...),
  ...
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted to each marginal distribution.
For instance, \code{formula = c(y1 + y2 + y3 ~ x1 + x2 | x1 + x3 | x2 + x3)} fits a
3-variate BCS-NSM regression that models \code{y1} as a function of \code{x1} and \code{x2};
\code{y2} as a function of \code{x1} and \code{x3}; and \code{y3} as a function of \code{x2}
and \code{x3}. See details below.}

\item{data}{an optional data frame containing the variables in the formula. By default the
variables are taken from environment(formula).}

\item{subset}{an optional vector specifying a subset of index observations to be used in the fitting
process. (See additional details about how this argument interacts with data-dependent bases
in the ‘Details’ section of the \code{\link[stats]{model.frame}} documentation.)}

\item{na.action}{a function which indicates what should happen when the data contain \code{NAs}.
The default is set by the \code{na.action} setting of \link[base]{options}, and is
\code{\link[stats]{na.fail}} if that is unset. The ‘factory-fresh’ default is
\code{\link[stats]{na.omit}}. Another possible value is \code{NULL}, no action.
Value \code{\link[stats]{na.exclude}} can be useful.}

\item{margins}{a character or a character vector; specifies the marginal BCS distributions. If all
BCS margins are the same, it is sufficient to enter only one character. A table with the
current available BCS distributions can be seen in \code{\link{bcs}}.}

\item{links}{a character or a character vector; specifies the link functions for the \code{mu}
regression structure. The links \code{"log"} (default) and \code{"identity"} are currently
available. If all links are the same, it is sufficient to enter only one character.}

\item{association}{one of \code{"unstructured"} (default), \code{"uniform"}, or
\code{"nonassociative"}, which specify the association matrix of the model.}

\item{copula}{character; informs which normal scale mixture distribution
should be used to generate the NSM copula. Currently,
the copulas available are: Gaussian (\code{"gaussian"}),
Student's \emph{t} (\code{"t"}), slash (\code{"slash"}), and hyperbolic (\code{"hyp"}).}

\item{eta}{possible extra parameter induced by the copula, parameterized to take values at
\code{(0, 1)}. For instance, under the \emph{t} copula, \code{eta} specifies
\code{eta} / (1 - \code{eta}) degrees of freedom. To estimate the value of \code{eta}
via profile log-likelihood, use the function \code{\link{choose_copula}}.}

\item{control}{a list of optimization control arguments specified via \code{\link{control_fit}}.}

\item{...}{further optimization control arguments passed to \code{\link{control_fit}}.}
}
\value{
The \code{bcsnsmreg} function returns an object of class "\code{bcsnsmreg}",
    which consists of a list with the following components:
\describe{
  \item{coefficients}{a list with the estimated regression coefficients for each marginal distribution.}
  \item{fitted.values}{the fitted medians for each marginal distribution.}
  \item{marginal.pars}{a list with the fitted values for the marginal parameters \code{mu}, \code{sigma},
      \code{lambda}, and \code{nu}.}
  \item{margins}{a character vector with the specified marginal BCS distributions.}
  \item{links}{a character vector with the specified link functions.}
  \item{copula, eta}{\code{"copula"} is a character which informs which normal scale mixture distribution
      was used to generate the NSM copula and \code{"eta"} is the possible extra parameter associated with
      the copula, parameterized to take values at (0, 1).}
  \item{gamma}{the estimated parameters of the association matrix, if any.}
  \item{association}{structure of the association matrix. It can be one of \code{"non-associative"},
      \code{"unstructured"}, or \code{"uniform"}.}
  \item{logLik}{log-likelihood of the fitted model.}
  \item{vcov}{asymptotic covariance matrix of the maximum likelihood estimator of the model parameters vector.
      Specifically, the inverse of the observed information matrix, obtained via numeric Hessian matrix.}
  \item{optim_params}{control optimization parameters used by \code{\link{control_fit}}.}
  \item{y}{the response matrix used.}
  \item{nobs,d}{the number of observations in the sample and the dimension of the response variable, respectively.}
  \item{call}{ the function call.}
  \item{formula}{the formula used to specify the model in \code{bcsnsm}.}
  \item{terms}{a list with the terms for the "\code{mu}" submodel for each marginal.}
 }
}
\description{
Fit a multivariate BCS-NSM regression via maximum likelihood for positive continuous
    data. Specify the marginal distributions within the class of Box-Cox symmetric (BCS) distributions
    (see \code{\link{bcs}}) and the dependence structure regarding the copula of a normal scale
    mixture distribution (see \code{\link{bcsnsm.dist}}).
}
\details{
The \code{formula} argument of the \code{formula} method uses functionalities inherited from
     package (Zeileis and Croissant, 2010). A basic formula for specifying a 3-variate BCS-NSM
     regression model is, for instance, \code{y1 + y2 + y3 ~ x1 + x2 + x3}, which specifies the
     same regression structure for the three marginal distributions as a function of the
     explanatory variables \code{x1, x2} and \code{x3}. However, using the Formula package
     allows different regression structures to be specified for each marginal distribution using
     the "\code{|}" operator. For instance, \code{formula = c(y1 + y2 + y3 ~ x1 + x2 | x1 + x3 |
     x2 + x3)} fits a 3-variate BCS-NSM regression that models \code{y1} as a function of
     \code{x1} and \code{x2}; \code{y2} as a function of \code{x1} and \code{x3}; and \code{y3}
     as a function of \code{x2} and \code{x3}.
      
 Parameter estimation is performed by directly maximizing the log-likelihood function through
     the \code{\link[stats]{optim}} function. To select the possible extra parameter (\code{eta})
     of the NSM copula, the \code{\link{choose_copula}} function can be used.
}
\examples{
\dontrun{

# Winscosin Breast Cancer Dataset
?wdbc

# Training set index
set.seed(123)
id <- sample(1:nrow(wdbc), 0.7 * nrow(wdbc))

# Reference model
fit0 <- bcsnsmreg(texture + area + smoothness + compactness + concavity ~ diagnosis,
                 data = wdbc, subset = id)
fit0

## Marginal quantile residuals of the reference model
plot(fit0, "marginal", panel = c(2, 3))

# Improve the fit on margins
fit_gaussian <- bcsnsmreg(texture + area + smoothness + compactness + concavity ~ diagnosis,
                         data = wdbc, subset = id, margins = c("lt", "lt", "lno", "lpe", "bct"))

## Marginal quantile residuals of the improved model
plot(fit_gaussian, "marginal", panel = c(2, 3))

## Summary
summary(fit_gaussian)

## Overall quantile residuals of the final model
plot(fit_gaussian)

## The epsilon's transformations
plot(fit_gaussian, "epsilon")
}

}
\references{
Ferrari, S. L., and Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. \emph{AStA Advances in Statistical Analysis}, 101, 321-344.

 Zeileis A., and Croissant, Y. (2010). Extended model formulas in
 R: multiple parts and multiple responses. \emph{Journal of Statistical Software},
 \bold{34}, 1–13.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
