% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_choose_copula.R
\name{choose_copula}
\alias{choose_copula}
\alias{print.choose_copula}
\alias{plot.choose_copula}
\title{Select the Extra Parameter for a Multivariate BCS-NSM Regression Fit}
\usage{
choose_copula(
  object,
  grid = seq(0.5, 0.98, 0.04),
  copula,
  trace = TRUE,
  plot = TRUE,
  control = control_fit(...),
  ...
)

\method{print}{choose_copula}(x, ...)

\method{plot}{choose_copula}(x, ...)
}
\arguments{
\item{object}{an object of class \code{"bcsnsm"}, a result of a call to \code{\link{bcsnsm}}.}

\item{grid}{numeric vector with values in (0, 1); the grid of \code{eta} values used to 
evaluate the profiled log-likelihood. If the tweaks are not computationally intensive, 
we suggest \code{grid = seq(0.5, 0.98, 0.04)}.}

\item{copula}{character; informs which normal scale mixture distribution
should be used to generate the NSM copula. The copulas with extra parameters are 
the \eqn{t} copula (\code{copula = "t"}), slash copula (\code{copula = "slash"}),
and hyperbolic copula (\code{copula = "hyp"}).}

\item{trace}{logical; if \code{TRUE}, a summary with the profiled log-likelihood value, the AIC,
and the BIC of the fit is displayed.}

\item{plot}{logical; if \code{TRUE}, a graph of the profiled log-likelihood evaluated in the
considered grid of values is shown.}

\item{control}{a list of control arguments specified via \code{\link[bcsnsm]{control_fit}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{"choose_copula"}.}
}
\value{
An object of class \code{"choose_copula"}, which is a named list where
    which each element is a \code{"\link{bcsnsm}"} fitted with a different value of the extra 
    parameter specified in \code{grid}.

    The \code{print} method summarizes the fits by displaying, for each value in \code{grid},
    the log-likelihood value and the Akaike (AIC) and Bayesian (BIC) information criteria. The
    \code{plot} method returns a graph of the profiled likelihood of the extra parameter, 
    .highlighting the value that maximizes the profiled log-likelihood.
}
\description{
Estimates the extra parameter in a multivariate BCS-NSM regression fit 
by profiling the log-likelihood over a grid of values in the interval (0, 1).
}
\details{
The extra parameter is denoted by \code{eta}, which is reparameterized to take values 
in the interval (0, 1) for numerical stability. For instance, if
\code{delta} is a positive extra parameter induced by the normal scale mixture copula,
we consider the reparameterization \code{eta = delta / (1 + delta)}. Thus, for example,
under the \emph{t} copula, \code{eta} corresponds to \code{eta / (1 - eta)} degrees of freedom.
This function evaluates the profiled log-likelihood over a grid of \code{eta} values and returns 
the fitted models for each one. The optimal value can be selected based on the log-likelihood, AIC, or BIC.
}
\examples{
# Data set: macronutrients (for description run ?macronutrients)
# Consider modeling the animal and plant protein regarding income level

## Response distribution:
mvplot(macronutrients[, 1:2])

## Animal and plant protein by income level:
boxplot(animal_p ~ income, macronutrients)
boxplot(plant_p ~ income, macronutrients)

## Reference model
fit0 <- bcsnsm(animal_p + plant_p ~ income | income, data = macronutrients)
fit0

## Fitting a t copula
fit_t <- choose_copula(fit0, copula = "t")

# Class
class(fit_t)

# It is possible to recover the plot:
plot(fit_t)

# And the model selection summary:
fit_t

# Fit with eta = 0.9
fit_t[["0.9"]]
summary(fit_t[["0.9"]])
}
\references{
Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and applications to
    nutritional data. \emph{AStA Advances in Statistical Analysis}, \bold{101}, 321--344.

    Medeiros, R. M. R. de, and Ferrari, S. L. P. (2025). Multivariate Box-Cox symmetric distributions
        generated by a normal scale mixture copula.
    
    Vanegas, L. H., and Paula, G. A. (2016). Log-symmetric distributions: statistical properties and
    parameter estimation. \emph{Brazilian Journal of Probability and Statistics}, \bold{30}, 196--220.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
