% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_bcsnsm_methods.R
\name{residuals.bcsnsm}
\alias{residuals.bcsnsm}
\title{Extract Model Residuals for a BerG Regression}
\usage{
\method{residuals}{bcsnsm}(object, type = c("mahalanobis", "epsilon"), ...)
}
\arguments{
\item{object}{an \code{"bcsnsm"} object.}

\item{type}{character; specifies which residual should be extracted. The available arguments are
\code{"mahalanobis"} (default), and \code{"epsilon"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract Model Residuals for a BerG Regression
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
