% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_methods.R
\name{summary.bcsnsm}
\alias{summary.bcsnsm}
\alias{print.summary.bcsnsm}
\title{Summarizing a BCS-NSM Multivariate Regression Fit}
\usage{
\method{summary}{bcsnsm}(object, ...)

\method{print}{summary.bcsnsm}(x, digits = max(3, getOption("digits") - 3), margin, ...)
}
\arguments{
\item{object}{an object of class \code{"bcsnsm"}, a result of a call to \code{\link{bcsnsm}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{"bcsnsm.sdlrm"}, a result of a call to \code{summary.bcsnsm}.}

\item{digits}{a non-null value for digits specifies the minimum number of significant digits to
be printed in values.}

\item{margin}{an integer value specifying a marginal for which the summary will
be presented. If missing (default), a summary for all marginal distributions 
is considered.}
}
\value{
The function \code{summary.bcsnsm} returns an object of class \code{"summary.bcsnsm"},
    which consists of a list with the following components:
 \describe{
    \item{call}{the original function call, given in \code{object}.}
    \item{beta_tab}{a list with summary statistics for the regression coefficients.}
    \item{sigma_tab}{summary statistics for the \code{sigma} marginal parameter vector.}
    \item{lambda_tab}{summary statistics for the \code{lambda} marginal parameter vector, if any.}
    \item{nu_tab}{summary statistics for the \code{nu} marginal parameter vector, if any.}
    \item{Gamma}{the fitted association matrix.}
    \item{margins}{a character vector with the specified marginal BCS distributions.}
    \item{links}{a character vector with the specified link functions.}
    \item{association}{structure specified for the association matrix. It can be one of
        \code{"non-associative"}, \code{"unstructured"}, or \code{"uniform"}.}
    \item{copula,eta}{\code{copula} is a character which informs which normal scale mixture 
        distribution was used to generate the NSM copula and \code{eta} is the 
        possible extra parameter associated with the copula, parameterized to take 
        values at (0, 1).}
    \item{residuals}{the overall quantile residuals.}
    \item{logLik}{log-likelihood value of the fitted model.}
    \item{y}{the response matrix used.}
    \item{AIC, BIC}{Akaike and Bayesian information criteria.}
 }
}
\description{
\code{summary} method for class \code{"bcsnsm"}.
}
\examples{
# Data set: macronutrients (for description run ?macronutrients)
# Consider modeling the animal and plant protein regarding income level

## Response distribution:
mvplot(macronutrients[, 1:2])

## Animal and plant protein by income level:
boxplot(animal_p ~ income, macronutrients)
boxplot(plant_p ~ income, macronutrients)

## Reference model
fit0 <- bcsnsm(animal_p + plant_p ~ income | income, data = macronutrients)
summary(fit0)

## Improved model
fit <- bcsnsm(animal_p + plant_p ~ income | income, data = macronutrients,
              copula = "t", eta = 0.9, margins = c("bcloii", "bct"))
summary(fit)               

## Summary with a specific margin
print(summary(fit), margin = 1)
print(summary(fit), margin = 2)
}
\references{
Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. \emph{AStA Advances in Statistical Analysis}, \bold{101}, 321--344.
     
 Medeiros, R. M. R., and Ferrari, S. L. P. (2025). Multivariate Box-Cox symmetric regression 
     models generated by a normal scale mixture copula.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
