% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_methods.R
\name{bcsnsm-methods}
\alias{bcsnsm-methods}
\alias{model.frame.bcsnsm}
\alias{model.matrix.bcsnsm}
\alias{coef.bcsnsm}
\alias{vcov.bcsnsm}
\alias{logLik.bcsnsm}
\alias{AIC.bcsnsm}
\title{Methods for BCS-NSM Multivariate Regression Fits}
\usage{
\method{model.frame}{bcsnsm}(formula, ...)

\method{model.matrix}{bcsnsm}(object, ...)

\method{coef}{bcsnsm}(object, ...)

\method{vcov}{bcsnsm}(object, parm = c("beta", "sigma", "lambda", "nu", "gamma", "all"), ...)

\method{logLik}{bcsnsm}(object, ...)

\method{AIC}{bcsnsm}(object, ..., k = 2)
}
\arguments{
\item{formula}{a model \link{formula} or \link{terms} object or an \code{"sdlrm"} object.}

\item{...}{additional argument(s) for methods.}

\item{object}{an object of class \code{"bcsnsm"}, a result of a call to \code{\link{bcsnsm}}.}

\item{parm}{a character; specifies which set of parameters should be used to extract the
asymptotic covariance matrix of the maximum likelihood estimators (for \code{vcov} function). The options are 
\code{"beta"} (default), \code{"sigma"}, \code{"lambda"}, \code{"nu"}, 
\code{"gamma"}, and \code{"all"}, where all parameters are considered}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical Akaike information criteria (AIC).}
}
\value{
\itemize{
\item \code{model.frame} returns a \code{data.frame} containing the variables required
    by \code{formula} and any additional arguments provided via \code{...}.
\item \code{model.matrix} returns a list, where each element consists of the corresponding 
    design matrix of that marginal distribution used in the regression structure.
\item \code{coef} returns a list, each element of which is a numeric vector with the 
    estimated regression coefficients for each margin.
\item \code{vcov} returns the asymptotic covariance matrix of the regression coefficients,
    based on the \code{parm} argument.
\item \code{logLik} returns the log-likelihood value of the fitted model.
\item \code{AIC} returns a numeric value representing the generalized information 
    criterion, such as AIC or BIC, depending on the value of \code{k}.
    The default \code{k = 2} corresponds to AIC, while \code{k = log(n)} gives BIC.
}
}
\description{
These are S3 methods for extracting information from 
    fitted objects of class \code{"bcsnsm"}, which represent multivariate 
    regression models based on the BCS-NSM framework. The available methods 
    include functions to extract estimated coefficients, standard errors, 
    variance-covariance matrices, log-likelihood values, information criteria,
    and the original model call.
}
\examples{
# Data set: macronutrients (for description run ?macronutrients)
# Consider modeling the animal and plant protein regarding income level

## Response distribution:
mvplot(macronutrients[, 1:2])

## Animal and plant protein by income level:
boxplot(animal_p ~ income, macronutrients)
boxplot(plant_p ~ income, macronutrients)

## Fit the model with Gaussian copula and BCNO margins
fit <- bcsnsm(animal_p + plant_p ~ income | income, data = macronutrients)

# Coef
coef(fit)

# vcov
vcov(fit)
vcov(fit, parm = "sigma")
vcov(fit, parm = "lambda")
vcov(fit, parm = "gamma")
vcov(fit, parm = "all")

# Log-likelihood value
logLik(fit)

# AIC and BIC
AIC(fit)
AIC(fit, k = log(fit$nobs))

# Model matrices
model.matrix(fit)      ## Model matrices for all margins
model.matrix(fit)[[1]] ## Model matrix for the first margin
model.matrix(fit)[[2]] ## Model matrix for the second margin
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
