% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_fit_functions.R
\name{bcsnsm}
\alias{bcsnsm}
\title{BCS-NSM Fit for Multivariate Positive Data}
\usage{
bcsnsm(
  y,
  margins = "bcno",
  association = c("unstructured", "uniform", "nonassociative"),
  copula = c("gaussian", "t", "slash", "hyp"),
  eta = NULL,
  control = control_fit(...),
  ...
)
}
\arguments{
\item{y}{a matrix with the sample observations.}

\item{margins}{a character or a character vector; specifies the marginal BCS distributions.
If all BCS margins are the same, it is sufficient to enter only one character. A table with
the current available BCS distributions can be seen in \code{\link{bcs}}.}

\item{association}{one of \code{"unstructured"} (default), \code{"uniform"}, or
\code{"nonassociative"}, which specify the association matrix of the model.}

\item{copula}{character; informs which normal scale mixture distribution should be used to generate
the NSM copula. Currently, the copulas available are: Gaussian (\code{"gaussian"}),
Student's \emph{t} (\code{"t"}), slash (\code{"slash"}), and hyperbolic (\code{"hyp"}).}

\item{eta}{possible extra parameter induced by the copula, parameterized to take values at
\code{(0, 1)}. For instance, under the \emph{t} copula, \code{eta} specifies
\code{eta} / (1 - \code{eta}) degrees of freedom. To estimate the value of \code{eta}
via profile log-likelihood, use the function \code{\link{choose_copula}}.}

\item{control}{a list of control arguments specified via \code{\link{control_fit}}.}

\item{...}{further arguments passed to \code{\link{control_fit}}.}
}
\value{
The \code{bcsnsm} function returns an object of class \code{"bcsnsm"},
    which consists of a list with the following components:
\describe{
  \item{mu, sigma, lambda, nu}{vectors with the estimated values for the marginal parameters
      \code{mu}, \code{sigma}, \code{lambda}, and \code{nu}, respectively.}
  \item{gamma}{the estimated parameters of the association matrix, if any.}
  \item{margins}{a character vector with the specified marginal BCS distributions.}
  \item{association}{structure of the association matrix. It can be one of \code{"non-associative"},
      \code{"unstructured"}, or \code{"uniform"}.}
  \item{copula, eta}{\code{"copula"} is a character which informs which normal scale mixture distribution
      was used to generate the NSM copula and \code{"eta"} is the possible extra parameter associated with
      the copula.}
  \item{logLik}{log-likelihood of the fitted model.}
  \item{vcov}{asymptotic covariance matrix of the maximum likelihood estimator of the model parameters vector.
      Specifically, the inverse of the observed information matrix, obtained via numeric Hessian matrix.}
  \item{y}{the response matrix.}
  \item{optim_pars}{control optimization parameters used by \code{\link{control_fit}}.}
  \item{nobs,d}{the number of observations in the sample and the dimension of the response variable, respectively.}
  \item{call}{ the function call.}
 }
}
\description{
Fit a multivariate BCS-NSM distribution via maximum likelihood for positive continuous
    data. Specify the marginal distributions within the class of Box-Cox symmetric (BCS) distributions
    (see \code{\link{bcs}}) and the dependence structure regarding the copula of a normal scale
    mixture distribution (see \code{\link{bcsnsm.dist}}).
}
\examples{
\dontrun{
# Winscosin Breast Cancer Dataset
?wdbc

# Training set index
set.seed(123)
id <- sample(1:nrow(wdbc), 0.7 * nrow(wdbc))

# Reference model
fit0 <- bcsnsmreg(texture + area + smoothness + compactness + concavity ~ diagnosis,
                  data = wdbc[id, ])
fit0

## Marginal quantile residuals of the reference model
plot(fit0, "marginal", panel = c(2, 3))

# Improve the fit on margins
fit_gaussian <- bcsnsmreg(texture + area + smoothness + compactness + concavity ~ diagnosis,
                          data = wdbc[id, ], margins = c("lt", "lt", "lno", "lpe", "bct"))

## Marginal quantile residuals of the improved model
plot(fit_gaussian, "marginal", panel = c(2, 3))


## Summary
summary(fit_gaussian)

## Overall quantile residuals of the final model
plot(fit_gaussian)

## The epsilon's transformations
plot(fit_gaussian, "epsilon")
}
}
\references{
Vanegas, L. H., and Paula, G. A. (2016). Log-symmetric distributions: statistical properties and
    parameter estimation. \emph{Brazilian Journal of Probability and Statistics}, 30, 196-220.

Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and applications to
    nutritional data. \emph{AStA Advances in Statistical Analysis}, 101, 321-344.

Medeiros, R. M. R. de, and Ferrari, S. L. P. (2023). Multivariate Box-Cox symmetric distributions
    generated by a normal scale mixture copula.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
