% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_fit_function.R
\name{bcsnsm}
\alias{bcsnsm}
\alias{print.bcsnsm}
\title{Multivariate BCS-NSM Regression}
\usage{
bcsnsm(
  formula,
  data,
  subset,
  na.action,
  margins = "bcno",
  links = "log",
  association = c("unstructured", "uniform", "nonassociative"),
  copula = c("gaussian", "t", "slash", "hyp"),
  eta = NULL,
  control = control_fit(...),
  ...
)

\method{print}{bcsnsm}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted to each marginal distribution.
For instance, \code{formula = c(y1 + y2 + y3 ~ x1 + x2 | x1 + x3 | x2 + x3)} fits a
3-variate BCS-NSM regression that models \code{y1} as a function of \code{x1} and \code{x2};
\code{y2} as a function of \code{x1} and \code{x3}; and \code{y3} as a function of \code{x2}
and \code{x3}. See details below.}

\item{data}{an optional data frame containing the variables in the formula. By default the
variables are taken from environment(formula).}

\item{subset}{an optional vector specifying a subset of index observations to be used in the fitting
process. (See additional details about how this argument interacts with data-dependent bases
in the ‘Details’ section of the \code{\link[stats]{model.frame}} documentation.)}

\item{na.action}{a function which indicates what should happen when the data contain \code{NAs}.
The default is set by the \code{na.action} setting of \link[base]{options}, and is
\code{\link[stats]{na.fail}} if that is unset. The ‘factory-fresh’ default is
\code{\link[stats]{na.omit}}. Another possible value is \code{NULL}, no action.
Value \code{\link[stats]{na.exclude}} can be useful.}

\item{margins}{a character or a character vector; specifies the marginal BCS distributions. If all
BCS margins are the same, it is sufficient to enter only one character. A table with the
current available BCS distributions can be seen in \code{\link{bcs}}.}

\item{links}{a character or a character vector; specifies the link functions for the \code{mu}
regression structure. The links \code{"log"} (default) and \code{"identity"} are currently
available. If all links are the same, it is sufficient to enter only one character.}

\item{association}{one of \code{"unstructured"} (default), \code{"uniform"}, or
\code{"nonassociative"}, which specify the association matrix of the model.}

\item{copula}{character; informs which normal scale mixture distribution
should be used to generate the NSM copula. Currently,
the copulas available are: Gaussian (\code{"gaussian"}),
Student's \emph{t} (\code{"t"}), slash (\code{"slash"}), and hyperbolic (\code{"hyp"}).}

\item{eta}{possible extra parameter induced by the copula, parameterized to take values at
\code{(0, 1)}. For instance, under the \emph{t} copula, \code{eta} specifies
\code{eta} / (1 - \code{eta}) degrees of freedom. To estimate the value of \code{eta}
via profile log-likelihood, use the function \code{\link{choose_copula}}.}

\item{control}{a list of optimization control arguments specified via \code{\link{control_fit}}.}

\item{...}{further optimization control arguments passed to \code{\link{control_fit}}.}

\item{x}{a fitted model object of class \code{"bcsnsm"}.}

\item{digits}{a non-null value for digits specifies the minimum number of significant digits to
be printed in values.}
}
\value{
The \code{bcsnsm} function returns an object of class "\code{bcsnsm}",
    which consists of a list with the following components:
\describe{
  \item{coefficients}{a list with the estimated regression coefficients for each marginal distribution.}
  \item{fitted.values}{the fitted medians for each marginal distribution.}
  \item{marginal.pars}{a list with the fitted values for the marginal parameters \code{mu}, \code{sigma},
      \code{lambda}, and \code{nu}.}
  \item{margins}{a character vector with the specified marginal BCS distributions.}
  \item{links}{a character vector with the specified link functions.}
  \item{copula, eta}{\code{"copula"} is a character which informs which normal scale mixture distribution
      was used to generate the NSM copula and \code{"eta"} is the possible extra parameter associated with
      the copula, parameterized to take values at (0, 1).}
  \item{gamma}{the estimated parameters of the association matrix, if any.}
  \item{association}{structure of the association matrix. It can be one of \code{"non-associative"},
      \code{"unstructured"}, or \code{"uniform"}.}
  \item{logLik}{log-likelihood of the fitted model.}
  \item{vcov}{asymptotic covariance matrix of the maximum likelihood estimator of the model parameters vector.
      Specifically, the inverse of the observed information matrix, obtained via numeric Hessian matrix.}
  \item{optim_params}{control optimization parameters used by \code{\link{control_fit}}.}
  \item{y}{the response matrix used.}
  \item{nobs,d}{the number of observations in the sample and the dimension of the response variable, respectively.}
  \item{call}{ the function call.}
  \item{formula}{the formula used to specify the model in \code{bcsnsm}.}
  \item{terms}{a list with the terms for the "\code{mu}" submodel for each marginal.}
 }
}
\description{
Fits a multivariate regression model for positive continuous data using the BCS-NSM class.
    The marginal distributions are chosen from the Box-Cox symmetric (BCS) family 
    (see \code{\link{bcs}}), and the dependence structure is modeled via a normal scale mixture
    (NSM) copula.
}
\details{
Let \eqn{\bold{Y}_1, \ldots, \bold{Y}_n} be \eqn{n} independent positive continuous random 
    vectors, where each \eqn{\bold{Y}_i = (Y_{i1}, \ldots, Y_{id})^\top} follows a BCS-NSM 
    distribution with marginal parameters \eqn{\bold{\mu}_i = (\mu_{i1}, \ldots, \mu_{id})^\top},
    \eqn{\bold{\sigma} = (\sigma_1, \ldots, \sigma_d)^\top}, \eqn{\bold{\lambda} = 
    (\lambda_1, \ldots, \lambda_d)^\top}; marginal density generating functions 
    \eqn{\bold{r}(u) = (r_1(u), \ldots, r_d(u))^\top}, \eqn{u \geqslant 0}, and NSM copula 
    generated by the mixing CDF \eqn{H} with association matrix \eqn{\bold{\Gamma}}. 
    The class of the multivariate BCS-NSM regression models is defined as
    
    \eqn{
    \begin{array}{rl}
    \bold{Y}_i  & \sim \textrm{BCS--NSM}_d(\bold{\mu}_i, \bold{\sigma}, \bold{\lambda}, \bold{\Gamma}; \bold{r}, H),\\
    \bold{\mu}_i & = \left(g_1^{-1}(\bold{x}_{i1}^\top \bold{\beta}_1), \ldots, g_d^{-1}(\bold{x}_{id}^\top \bold{\beta}_d)\right)^\top, 
    \end{array}
    }
    
    where \eqn{\bold{\beta}_j} is a \eqn{b_j \times 1} vector of unknown regression coefficients; 
    \eqn{\bold{x}_{ij}} is a \eqn{b_j \times 1} vector with the observations on \eqn{b_j} 
    explanatory variables, which are assumed to be fixed and known; and \eqn{g_j : \mathbb{R}_+ 
    \longrightarrow \mathbb{R}} is a strictly monotone and at least twice differentiable link function.
    
 The \code{formula} argument uses functionalities inherited from \code{Formula}
     package (Zeileis and Croissant, 2010). A basic formula for specifying a 3-variate BCS-NSM
     regression model is, for instance, \code{y1 + y2 + y3 ~ x1 + x2 + x3}, which specifies the
     same regression structure for the three marginal distributions as a function of the
     explanatory variables \code{x1, x2} and \code{x3}. However, using the Formula package
     allows different regression structures to be specified for each marginal distribution using
     the "\code{|}" operator. For instance, \code{formula = c(y1 + y2 + y3 ~ x1 + x2 | x1 + x3 |
     x2 + x3)} fits a 3-variate BCS-NSM regression that models \code{y1} as a function of
     \code{x1} and \code{x2}; \code{y2} as a function of \code{x1} and \code{x3}; and \code{y3}
     as a function of \code{x2} and \code{x3}.
 
    The class of the BCS-NSM multivariate regression models  includes models with various 
    possibilities of marginal and dependence settings. For instance, each margin can follow a 
    different BCS distribution. This property provides the flexibility to accommodate marginal 
    distributions with different behaviors, such as different types of skewness and tail-heaviness.

    The NSM copula is characterized by the mixing cumulative distribution function (CDF) \eqn{H}. 
    The \code{bcsnsm} function provides the Gaussian, \eqn{t}, slash, and hyperbolic copulas, 
    specified by the \code{copula} argument. Note that some NSM copulas also have an extra 
    parameter, say \eqn{\delta}, which is assumed to be known and fixed in the BCS-NSM 
    multivariate regression. This parameter is generally positive, but we consider the 
    parameterization \eqn{\eta = \delta / (1 + \delta) \in (0, 1)} to make it easier to choose 
    its value in practice.
    
    The association matrix \eqn{\Gamma} plays an important role in describing the dependence.
    Let \eqn{\gamma_{ij}} be the element of the \emph{i}th row and \emph{j}th column of the 
    association matrix \eqn{\bold{\Gamma}}, where \eqn{\gamma_{ii} = 1}, for 
    \emph{i, j = 1, ..., d}. Thus, Kendall's tau of the \eqn{i}th and \eqn{j}th margins is given by

    \deqn{\tau_{ij} = \dfrac{2 \; \textrm{asin}(\gamma_{ij})}{\pi},}

    and it is invariant under the copula. Since Kendall’s tau is a rank correlation coefficient, 
    it measures any monotonous association between two random variables. 
      
 Parameter estimation is performed by directly maximizing the log-likelihood function through
     the \code{\link[stats]{optim}} function. The \code{\link{choose_copula}} function can be used
     to select the possible extra parameter (\code{eta}) of the NSM copula.
}
\examples{
\donttest{#' # Data set: macronutrients (for description run ?macronutrients)
# Consider modeling the animal and plant protein regarding income level

## Response distribution:
mvplot(macronutrients[, 1:2])

## Animal and plant protein by income level:
boxplot(animal_p ~ income, macronutrients)
boxplot(plant_p ~ income, macronutrients)

## Step 1: Reference model
fit0 <- bcsnsm(animal_p + plant_p ~ income | income, data = macronutrients)
fit0

## Marginal quantile residuals of the reference model
plot(fit0, "marginal", panel = c(2, 3))

## Step 2: Improve the fit on margins
fit_gaussian <- bcsnsm(animal_p + plant_p ~ income | income, data = macronutrients,
                       margins = c("bcloii", "bct"))

## Marginal quantile residuals of the improved model
plot(fit_gaussian, "marginal", panel = c(2, 3))

## Step 3: Choose a copula

### Choosing the extra parameter of the t copula (it can be slow)
fit_t <- choose_copula(fit_gaussian, copula = "t")
plot(fit_t)

## Comparing with AIC
AIC(fit_gaussian)
AIC(fit_t[["0.9"]])

## Overall quantile residuals
plot(fit_gaussian)
plot(fit_t[["0.9"]])

## The epsilon's transformations
plot(fit_gaussian, "epsilon")
plot(fit_t[["0.9"]], "epsilon")}
}
\references{
Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. \emph{AStA Advances in Statistical Analysis}, \bold{101}, 321--344.
     
 Medeiros, R. M. R., and Ferrari, S. L. P. (2025). Multivariate Box-Cox symmetric regression 
     models generated by a normal scale mixture copula.

 Zeileis A., and Croissant, Y. (2010). Extended model formulas in
 R: multiple parts and multiple responses. \emph{Journal of Statistical Software},
 \bold{34}, 1–-13.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
