% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_bcsnsm_methods.R
\name{plot.bcsnsm}
\alias{plot.bcsnsm}
\title{Visualization of the fit of the BCNSM distributions}
\usage{
\method{plot}{bcsnsm}(
  x,
  type = c("response", "marginal", "epsilon"),
  outliers = FALSE,
  alpha = 0.01,
  panel = NULL,
  levels = c(1, 0.1, 0.01, 0.001, 1e-04),
  ...
)
}
\arguments{
\item{x}{an object of class \code{bcsnsm}.}

\item{type}{character; specifies which graphical should be produced. The available options
are \code{"response"} (default), \code{"margins"}, and \code{"epsilon"}.}

\item{outliers}{logical; used only when \code{type = "response"}. If \code{TRUE},
possible outliers are highlighted in red.}

\item{alpha}{criterion according to the squared Mahalanobis distances that identifies a point as
a possible outlier.}

\item{panel}{A vector of the form \code{c(nr, nc)} with the number of rows and columns, where
the figures will be drawn in an \code{nr-}by\code{-nc} array on the device.}

\item{levels}{levels for contours plots.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Visualization of the fit of the BCNSM distributions
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
