% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_methods.R
\name{plot.bcsnsm}
\alias{plot.bcsnsm}
\title{Diagnostic Plots for a Multivariate BCS-NSM Regression Fit}
\usage{
\method{plot}{bcsnsm}(
  x,
  type = c("quantile", "marginal", "mahalanobis", "epsilon"),
  levels = c(0.1, 0.01, 0.001, 1e-04),
  panel = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"bcsnsm"}, a result of a call to \code{\link{bcsnsm}}.}

\item{type}{character; specifies which graphical should be produced. The available options
are \code{"quantile"} (default), \code{"marginal"}, \code{"mahalanobis"}, and
\code{"epsilon"}. Se details below.}

\item{levels}{levels for contours plots. Used only when \code{type = "epsilon"}.}

\item{panel}{a vector of the form \code{c(nr, nc)} with the number of rows and columns, where
the figures will be drawn in an \code{nr-}by\code{-nc} array on the device. Used
only when \code{type = "marginal"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{plot} method for \code{"\link{bcsnsm}"} objects returns four types
    of diagnostic plots.
}
\description{
This function provides tools for evaluating the goodness-of-fit of a BCS-NSM
    regression to a multivariate positive data.
}
\details{
The argument \code{type} specifies what type of diagnostic plot should be produced. The available
    options are \code{"quantile"} (default), \code{"marginal"}, \code{"mahalanobis"}, and
    \code{"epsilon"}. When \code{type = "quantile"}, the function returns the normal probability
    plot with a 95\% confidence region of the overall quantile residuals, where the confidence
    region is computed according to Fox (2015, Sec. 3.1.3). 
    
    When \code{type = "marginal"}, it 
    produces the normal probability plots of the quantile residuals resulting from the fit at
    each marginal distribution. The argument \code{panel} is a vector \code{c(nr, nc)},
    indicating the number of rows and columns in which these graphs will be arranged on the 
    device. 
    
    When \code{type = "mahalanobis"}, a plot of the transformed Mahalanobis distances 
    for each observational unit is produced. Finally, when \code{type = "epsilon"}, a panel
    graphical summary of the epsilon transformations is returned, where the bottom panels show 
    the scatterplots with the fitted contour lines for
    each pair at \code{levels}; the diagonal panels show the histograms with the fitted marginal
    densities; and the top panels show the upper triangular elements of the fitted association matrix. 
    
    Details about the diagnostic tools used by the \code{plot.bcsnsm} function can be found in 
    \code{\link{residuals.bcsnsm}} and in Medeiros and Ferrari (2025).
}
\examples{
# Data set: macronutrients (for description run ?macronutrients)
# Consider modeling the animal and plant protein regarding income level

## Response distribution:
mvplot(macronutrients[, 1:2])

## Animal and plant protein by income level:
boxplot(animal_p ~ income, macronutrients)
boxplot(plant_p ~ income, macronutrients)

## Reference model
fit <- bcsnsm(animal_p + plant_p ~ income | income, data = macronutrients)

## Overall quantile residuals (default)
plot(fit)

## Marginal quantile residuals
plot(fit, type = "marginal")
plot(fit, type = "marginal", panel = c(2, 1)) # Change the layout

## Mahalanobis distances
plot(fit, type = "mahalanobis")

## The epsilon's transformations
plot(fit, type = "epsilon")
}
\references{
Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. \emph{AStA Advances in Statistical Analysis}, \bold{101}, 321--344.
     
 Fox, J. (2015). \emph{Applied Regression Analysis and Generalized Linear Models}. 
     Sage Publications.
     
 Medeiros, R. M. R., and Ferrari, S. L. P. (2025). Multivariate Box-Cox symmetric regression 
     models generated by a normal scale mixture copula.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
