% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_bcsnsm_methods.R
\name{residuals.bcsnsm}
\alias{residuals.bcsnsm}
\title{Extract Residuals for an \code{bcsnsm} Object}
\usage{
\method{residuals}{bcsnsm}(object, type = c("mahalanobis", "epsilon"), ...)
}
\arguments{
\item{object}{an \code{"bcsnsm"} object.}

\item{type}{character; specifies which residual should be extracted. The available arguments are
\code{"mahalanobis"} (default), and \code{"epsilon"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Two quantities that allow the diagnosis of the fit of a BCS-NSM distribution. The transformed 
    Mahalanobis distances make it possible to detect outliers and the epsilon transformations,
    \eqn{\varepsilon_{ij} = \Psi_H^{-1}\{\widehat{F}_{j}(y_{ij})\}}, alows to evaluate the
     marginal and copula assumptions.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
