% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_methods.R
\name{residuals.bcsnsm}
\alias{residuals.bcsnsm}
\title{Extract Residuals from BCS-NSM Multivariate Regression Fits}
\usage{
\method{residuals}{bcsnsm}(
  object,
  type = c("quantile", "marginal", "mahalanobis", "epsilon"),
  ...
)
}
\arguments{
\item{object}{an object of class \code{"bcsnsm"}, a result of a call to \code{\link{bcsnsm}}.}

\item{type}{character; specifies which residual should be extracted. The available options are
\code{"quantile"} (default), \code{"marginal"}, \code{"mahalanobis"}, and \code{"epsilon"}.
See details below.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
If \code{type = "quantile"} or \code{type = "mahalanobis"}, the \code{residual.bcsnsm}
    function returns a vector with the overall quantile residuals or the transformed Mahalanobis 
    (squared) distances, respectively. Conversely, if \code{type = "marginal"} or 
    \code{type = "epsilon"}, it returns a matrix with the quantile residuals or the epsilon 
    transforms of each marginal fit, respectively.
}
\description{
Extracts several types of residuals for assessing the goodness-of-fit of BCS-NSM multivariate 
    regression models. Available residuals include: (i) marginal quantile residuals for 
    evaluating the adequacy of each marginal distribution; (ii) epsilon transformations for 
    diagnostic analysis under the copula structure; (iii) squared Mahalanobis distances of the 
    epsilon transformations to detect multivariate outliers; and (iv) overall quantile residuals 
    based on the distribution of the Mahalanobis distances.
}
\details{
The fit of a BCS-NSM multivariate regression has some assumptions that need to be evaluated
    with a diagnostic analysis. Medeiros and Ferrari (2025), using the properties of normal
    scale mixture copulas, proposed diagnostic tools that can be used to assess the 
    goodness-of-fit in practical applications. 

The simplest is based on obtaining the quantile residuals (Dunn and Smith, 1996) for each 
    Box-Cox symmetric marginal distribution, which is useful for assessing goodness-of-fit at 
    the margins of the postulated model. These residuals are called marginal quantile residuals,
    and they can be accessed with the function \code{residuals.bcsnsm} with the 
    argument \code{type = "marginal"}. If the \eqn{j}th BCS marginal distribution is well-fitted
    to the model, the marginal quantile residuals are expected to have an approximately standard 
    normal distribution.

The other options available are based on the relationship that the BCS-NSM distributions have 
    with the normal scale mixtures (NSM) distributions. Let 
    \eqn{\bold{Y}_i = (Y_{i1}, \ldots,Y_{id})^\top}, for \eqn{i \in \{1, \ldots, n\}}, be a 
    sample of random vectors under the specification of a multivariate BCS-NSM regression, with 
    association matrix \eqn{\bold{\Gamma}}.

The BCS-NSM distributions are related to the NSM distributions through epsilon transformations:
  
  \eqn{
    \varepsilon_{ij} = \Psi_{H}^{-1}\left\{F_j(Y_{ij})\right\},
  }

where \eqn{\Psi_H} denotes the cumulative distribution function (CDF) of a standard univariate 
    NSM distribution and \eqn{F_j} denotes the CDF of the \eqn{j}th marginal BCS distribution. 
    The NSM distributions are a class of multivariate distributions whose elements are 
    characterized by the CDF \eqn{H}, called the mixing CDF. 

If \eqn{\bold{\varepsilon}_j = (\varepsilon_{1j}, \ldots, \varepsilon_{nj})^\top}, it can be 
    shown that the vectors \eqn{\bold{\varepsilon_{1}}, \ldots, \bold{\varepsilon_{d}}} form 
    a \eqn{d}-variate random sample of the NSM distribution with location vector 
    \eqn{(0, \ldots, 0)}, dispersion matrix \eqn{\bold{\Gamma}} (the same association matrix 
    as the corresponding BCS-NSM distribution), and mixing CDF \eqn{H}. 
  
  
  For instance, if a BCS-NSM regression with Gaussian copula is assumed, the epsilon 
      transformations are expected to have a multivariate normal distribution approximately. 
      On the other hand, if the \eqn{t} copula is assumed, the transformations are expected to 
      have a multivariate \eqn{t} distribution, and so on.
  
  The matrix with the epsilon transformations can be accessed with the function 
      \code{residuals.bcsnsm} with the argument \code{type = "epsilon"}. The squared 
      Mahalanobis distances of the epsilon transformations, i.e.,
  
  \eqn{
  M_i = \bold{\varepsilon}_i^\top \bold{\Gamma}^{-1} \bold{\varepsilon}_i, \quad i \in \{1, \ldots, n\}
  }

    are useful for identifying outliers in the multivariate context. The transformed Mahalanobis 
    distances can be accessed with the function \code{residuals.bcsnsm} with the argument 
    \code{type = "mahalanobis"}.

Furthermore, the CDF of the distances is known and can be written as (Lange and Sinsheimer, 1993)

\eqn{
  F_M(m) = P(M_i \leqslant m) = \dfrac{m^{d/2}}{2^{d/2} \Gamma(d/2)} \int_0^\infty H(1/u) u^{d/2-1} e^{-mu/2} \textrm{d}u, \quad m > 0.
}

This result allows us to define a general quantile residual to evaluate the overall 
    goodness-of-fit more simply than with the epsilon transformations. Medeiros and 
    Ferrari (2025) define the so-called overall quantile residuals as

\eqn{
  r_{i}^q = \Psi^{-1}[F_M(M_i)], \quad i \in \{1, \ldots, n\}
}

If the model is well-fit to the data, the overall quantile residuals are expected to have an 
    approximately standard normal distribution. The overall quantile residuals can be accessed 
    with the function \code{residuals.bcsnsm} with the argument \code{type = "quantile"} (default).
}
\examples{
# Data set: macronutrients (for description run ?macronutrients)
# Consider modeling the animal and plant protein regarding income level

## Response distribution:
mvplot(macronutrients[, 1:2])

## Animal and plant protein by income level:
boxplot(animal_p ~ income, macronutrients)
boxplot(plant_p ~ income, macronutrients)

## Fit the model with Gaussian copula and BCNO margins
fit <- bcsnsm(animal_p + plant_p ~ income | income, data = macronutrients)

## Overall quantile residuals (default)
rq <- residuals(fit)
qqnorm(rq)
qqline(rq)

## Marginal quantile residuals
rm <- residuals(fit, "marginal")
qqnorm(rm[, 1]); qqline(rm[, 1])
qqnorm(rm[, 2]); qqline(rm[, 2])

## Transformed mahalanobis distances
maha <- residuals(fit, "mahalanobis")
plot(maha, type = "h")

## Epsilon transformations
epsilon <- residuals(fit, "epsilon")
plot(epsilon, pch = 16)
qqnorm(epsilon[, 1]); qqline(epsilon[, 1])
qqnorm(epsilon[, 2]); qqline(epsilon[, 2])
}
\references{
Dunn, P. K., and Smyth, G. K.(1996). Randomized quantile residuals. 
 \emph{Journal of Computational and Graphical Statistics}, \bold{5}, 236–-24

 Ferrari, S. L., and Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. \emph{AStA Advances in Statistical Analysis}, \bold{101}, 321--344.
     
 Lange, K., and Sinsheimer, J. S. (1993). Normal/independent distributions and their applications 
     in robust regression. \emph{Journal of Computational and Graphical Statistics}, \bold{2}, 175-–198      

 Medeiros, R. M. R., and Ferrari, S. L. P. (2025). Multivariate Box-Cox symmetric regression 
     models generated by a normal scale mixture copula.
 
 Zeileis A., and Croissant, Y. (2010). Extended model formulas in
 R: multiple parts and multiple responses. \emph{Journal of Statistical Software},
 \bold{34}, 1–13.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
