% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_regression_methods.R
\name{residuals.bcsnsmreg}
\alias{residuals.bcsnsmreg}
\title{Extract Residuals for an \code{bcsnsmreg} Object}
\usage{
\method{residuals}{bcsnsmreg}(
  object,
  type = c("quantile", "marginal", "mahalanobis", "epsilon"),
  ...
)
}
\arguments{
\item{object}{an \code{"bcsnsmreg"} object.}

\item{type}{character; specifies which residual should be extracted. The available arguments are
\code{"quantile"} (default), \code{"marginal"}, \code{"mahalanobis"}, and \code{"epsilon"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
It provides the overall and marginal quantile residuals, transformed Mahalanobis distances,
     and the epsilon transformations, \eqn{\varepsilon_{ij} = \Psi_H^{-1}\{\widehat{F}_{j}(y_{ij})\}}.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
